% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{expected_frequencies}
\alias{expected_frequencies}
\title{Produce ggplot of percentiles for simulated frequencies}
\usage{
expected_frequencies(r)
}
\arguments{
\item{r}{A dataframe. The result returned from \code{\link{sim_n_persons_x_times}}.}
}
\value{
Draws a ggplot
}
\description{
Produce ggplot of percentiles for simulated frequencies
}
\examples{
r <- sim_n_persons_x_times(dexp(1:30, rate = .05), n = 50, a = 5:7, ap = 1:3, 100)
expected_frequencies(r)

}
\seealso{
Other Utilities: \code{\link{calc_probabilities}},
  \code{\link{prob_categories}}
}

