% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsampler-package.R
\docType{package}
\name{gridsampler-package}
\alias{gridsampler-package}
\title{\pkg{gridsampler} contains tools to simulate the sampling of
multiple repertory grids.}
\description{
\pkg{gridsampler} contains tools to simulate the sampling of
multiple repertory grids.
}
\references{
TODO
}
\keyword{package}
\keyword{repgrid}

