% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsampler-package.R, R/gui.R
\docType{package}
\name{gridsampler}
\alias{gridsampler}
\alias{gridsampler-package}
\title{\pkg{gridsampler} contains tools to simulate the sampling of
multiple repertory grids.}
\usage{
gridsampler()
}
\description{
\pkg{gridsampler} contains tools to simulate the sampling of
multiple repertory grids.

Start gridsampler GUI
}
\examples{
\dontrun{
gridsampler()
} 

}
\author{
Mark Heckmann
}
\references{
TODO
}
\keyword{package}
\keyword{repgrid}

