% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_one_person}
\alias{sim_one_person}
\title{Simulate a single grid}
\usage{
sim_one_person(prob, a = 10)
}
\arguments{
\item{prob}{Probability to draw a construct from a certain category}

\item{a}{Number of constructs to be sampled}
}
\description{
Simulate a single grid
}
\examples{
# draw from exponential distribution
p <- dexp(1:20, rate=.1)
sim_one_person(p, a=10)

}
\keyword{internal}

