% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-utils.R
\name{text_to_vector}
\alias{text_to_vector}
\title{Convert a text input to a vector}
\usage{
text_to_vector(txt)
}
\arguments{
\item{txt}{Compatible text input via \link[shiny]{textInput}}
}
\value{
A vector
}
\description{
Convert a text input to a vector
}
\examples{
txt <- "10, 20, 30"
text_to_vector(txt)
}
\keyword{internal}

