% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsampler-package.R
\docType{package}
\name{gridsampler-package}
\alias{gridsampler-package}
\title{\pkg{gridsampler} - A sample size simulation software for repertory grid studies}
\description{
\pkg{gridsampler} - A sample size simulation software for repertory grid studies
}
\references{
\itemize{
 \item Green, B. (2004). Personal construct psychology and content analysis. Personal Construct Theory & Practice, 1(3), 82-91.
 \item Jankowicz, D. (2004). The easy guide to repertory grids. Chichester, England: John Wiley & Sons.
 }
}
\keyword{package}
\keyword{repgrid}

