% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsampler-package.R, R/gui.R, R/shiny-utils.R
\docType{package}
\name{gridsampler}
\alias{gridsampler}
\alias{gridsampler-package}
\title{\pkg{gridsampler} contains tools to simulate the sampling of
multiple repertory grids.}
\usage{
gridsampler()

gridsampler()
}
\value{
A shiny app
}
\description{
\pkg{gridsampler} contains tools to simulate the sampling of
multiple repertory grids.

Start gridsampler GUI

This function starts the gridsampler shiny app.
}
\examples{
\dontrun{
gridsampler()
} 

\dontrun{
gridsampler()
}
}
\author{
Mark Heckmann
}
\references{
TODO
}
\keyword{package}
\keyword{repgrid}

