% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridsampler-package.R, R/gui.R, R/shiny-utils.R
\docType{package}
\name{gridsampler}
\alias{gridsampler}
\alias{gridsampler-package}
\title{\pkg{gridsampler} contains tools to simulate the sampling of
multiple repertory grids.}
\usage{
gridsampler(display.mode = "auto",
  launch.browser = getOption("shiny.launch.browser", interactive()))

gridsampler(display.mode = "auto",
  launch.browser = getOption("shiny.launch.browser", interactive()))
}
\arguments{
\item{display.mode}{\code{auto} by default, can also be \code{showcase}.
See \link[shiny]{runApp}.}

\item{launch.browser}{Boolean, set \code{TRUE} to open the app in the browser.
See \link[shiny]{runApp}.}
}
\value{
A shiny app
}
\description{
\pkg{gridsampler} contains tools to simulate the sampling of
multiple repertory grids.

Start gridsampler GUI

This function starts the gridsampler shiny app.
}
\examples{
\dontrun{
gridsampler()
} 

\dontrun{
gridsampler()
}
}
\author{
Mark Heckmann
}
\references{
TODO
}
\keyword{package}
\keyword{repgrid}

