\docType{package}
\name{gridsampler}
\alias{gridsampler}
\alias{gridsampler-package}
\title{\pkg{gridsampler} contains tools to simulate the sampling of
multiple repertory grids.}
\description{
  \pkg{gridsampler} contains tools to simulate the sampling
  of multiple repertory grids.
}
\author{
  Mark Heckmann
}
\references{
  TODO
}
\keyword{package}
\keyword{repgrid}

