% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gridsampler}
\alias{gridsampler}
\title{Run gridsampler app}
\usage{
gridsampler(display.mode = "auto",
  launch.browser = getOption("shiny.launch.browser", interactive()))
}
\arguments{
\item{display.mode}{\code{auto} by default, can also be \code{showcase}.
See \link[shiny]{runApp}.}

\item{launch.browser}{Boolean, set \code{TRUE} to open the app in the browser.
See \link[shiny]{runApp}.}
}
\description{
This function starts the gridsampler shiny app.
}
\examples{
\dontrun{
gridsampler()
}
}

